Build Network of Public Clusters

Given node-level meta data for each sample's filtered public clusters, combine the data across samples and perform network analysis.

Usage

buildPublicClusterNetwork(

  ## Input ##
  file_list =
    list.files(file.path(getwd(), "public_clusters", "node_meta_data")),
  input_type = "rds", data_symbols = "ndat", header = TRUE, sep = "",
  seq_col,

  ## Network ##
  drop_isolated_nodes = FALSE,
  node_stats = NULL, stats_to_include = NULL, cluster_stats = NULL,

  ## Visualization ##
  color_nodes_by = "SampleID", color_scheme = "turbo",

  ## Output ##
  output_dir = file.path(getwd(), "public_clusters"),
  output_name = "PublicClusterNetwork",

  ...

)

Arguments

file_list

Passed to loadDataFromFileList when loading the node-level meta data for each sample (one data frame per sample).

input_type

Passed to loadDataFromFileList.

data_symbols

Passed to loadDataFromFileList.

header

Passed to loadDataFromFileList.

sep

Passed to loadDataFromFileList.

seq_col

The column name or number of each sample's data frame that contains the receptor sequences to be used as the basis of similarity between rows during network analysis. This column must have the same name in each sample.

drop_isolated_nodes

Passed to buildRepSeqNetwork.

node_stats

Deprecated. Does nothing. All node-level network properties are automatically computed.

stats_to_include

Deprecated. Does nothing.

cluster_stats

Deprecated. Does nothing. All cluster-level network properties are automatically computed.

color_nodes_by

Passed to buildRepSeqNetwork.

color_scheme

Passed to buildRepSeqNetwork.

output_dir

Passed to buildRepSeqNetwork.

output_name

Passed to buildRepSeqNetwork.

...

Other arguments to buildRepSeqNetwork.

Details

The input data is intended to be obtained using findPublicClusters() on the unfiltered RepSeq data for each sample.

Node-level network properties are renamed to reflect their association to the public network. The names used are ClusterIDPublic, PublicNetworkDegree, PublicTransitivity, PublicCloseness, PublicCentralityByCloseness, PublicEigenCentrality, PublicCentralityByEigen, PublicBetweenness, PublicCentralityByBetweenness, PublicAuthorityScore, PublicCoreness, PublicPageRank.

Value

A list of network objects as returned by buildRepSeqNetwork().

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

https://www.frontiersin.org/articles/10.3389/fimmu.2023.1181825

https://github.com/mlizhangx/Network-Analysis-for-Repertoire-Sequencing-

Author

Brian Neal (Brian.Neal@ucsf.edu)

See also

Examples

if (FALSE) {
vignette("public_clusters", package = "NAIR")
}