Build Public Cluster Network By Representative

Given cluster-level meta data for each sample's filtered public clusters, combine the data across samples and perform network analysis using a representative sequence from each cluster.

Usage

buildPublicClusterNetworkByRepresentative(

  ## Input ##
  file_list,
  input_type = "rds", data_symbols = "cdat",
  header = TRUE, sep = "",
  seq_col = "seq_w_max_count",
  count_col = "agg_count",

  ## Network ##
  dist_type = "hamming",
  dist_cutoff = 1,
  cluster_fun = cluster_fast_greedy,

  ## Visualization ##
  plots = TRUE,
  print_plots = TRUE,
  plot_title = "auto",
  plot_subtitle = "auto",
  color_nodes_by = "SampleID",
  color_scheme = "turbo",
  ...,

  ## Output ##
  output_dir = file.path(getwd(), "public_clusters"),
  output_type = "rda",
  output_name = "PubClustByRepresentative",
  pdf_width = 12, pdf_height = 10

)

Arguments

file_list

Passed to loadDataFromFileList when loading the cluster-level meta data for each sample (one data frame per sample). Data frames should match the format of the cluster-level meta data returned by buildRepSeqNetwork.

input_type

Passed to loadDataFromFileList.

data_symbols

Passed to loadDataFromFileList.

header

Passed to loadDataFromFileList.

sep

Passed to loadDataFromFileList.

seq_col

The column name or number of each data frame that contains the representative sequence to be used as the basis of similarity between rows during network analysis. This column must have the same name in each sample.

count_col

Passed to buildRepSeqNetwork.

dist_type

Passed to buildRepSeqNetwork.

dist_cutoff

Passed to buildRepSeqNetwork.

cluster_fun

Passed to addClusterMembership. Controls the algorithm used for cluster identification.

plots

Passed to buildRepSeqNetwork.

print_plots

Passed to buildRepSeqNetwork.

plot_title

Passed to buildRepSeqNetwork.

plot_subtitle

Passed to buildRepSeqNetwork.

color_nodes_by

Passed to generateNetworkGraphPlots.

color_scheme

Passed to generateNetworkGraphPlots.

...

Other arguments to generateNetworkGraphPlots.

output_dir

Passed to saveNetwork.

output_type

Passed to saveNetwork.

output_name

Passed to saveNetwork.

pdf_width

Passed to saveNetwork.

pdf_height

Passed to saveNetwork.

Details

This function is intended for performing network analysis on the public clusters obtained using findPublicClusters().

By using the cluster-level meta data as the input to buildRepSeqNetwork(), this function treats each public cluster as a single node and performs network analysis with similarity based on a representative sequence for each cluster (e.g., the sequence with the greatest clone count). buildRepSeqNetwork is called without any filtering and with drop_isolated_nodes = FALSE.

All node-level properties are automatically computed for the network, and cluster-level properties for the network (in which clusters of nodes represent clusters of public-cluster-representatives) are computed based on the properties of the public cluster represented by each node. See the 'value' section for a list and description of each property.

Value

A list of network objects as per buildRepSeqNetwork(). The data frame

cluster_data includes the following variables:

TotalSampleLevelNodes

Each network node in the cluster for the current row represents a public cluster; this value is the sum of the sample-network-level nodes contained in all such public clusters for the current row.

TotalCloneCount

Each network node in the cluster for the current row represents a public cluster; this value is the sum of the aggregate clone count across all such public clusters for the current row.

MeanOfMeanSeqLength

Each network node in the cluster for the current row represents a public cluster; this value is the mean value of the mean sequence length across all such public clusters for the current row.

MeanDegreeInPublicNet

The mean network degree of all the network nodes in the cluster for the current row.

MaxDegreeInPublicNet

The maximum network degree of all the network nodes in the cluster for the current row.

SeqWithMaxDegree

The representative sequence of the network node with maximum network degree of all the network nodes in the cluster for the current row. If more than one node attains the maximum network degree, the first sequence found is returned.

MaxCloneCount

Each network node in the cluster for the current row represents a public cluster; this is the maximum value of the maximum clone count property across all such public clusters for the current row.

SampleWithMaxCloneCount

The sample possessing the public cluster with the maximum value of the maximum clone count property across all the public clusters represented by the network nodes in the cluster for the current row. If more than one public cluster attains the maximum value, the first sample ID found is returned.

SeqWithMaxCloneCount

The representative sequence of the public cluster with the maximum value of the maximum clone count property across all the public clusters represented by the network nodes in the cluster for the current row. If more than one public cluster attains the maximum value, the first sequence found is returned.

MaxAggCloneCount

Each network node in the cluster for the current row represents a public cluster; this is the maximum value of the aggregate clone count property across all such public clusters for the current row.

SampleWithMaxAggCloneCount

The sample possessing the public cluster with the maximum value of the aggregate clone count property across all the public clusters represented by the network nodes in the cluster for the current row. If more than one public cluster attains the maximum value, the first sample ID found is returned.

SeqWithMaxAggCloneCount

The representative sequence of the public cluster with the maximum value of the aggregate clone count property across all the public clusters represented by the network nodes in the cluster for the current row. If more than one public cluster attains the maximum value, the first sequence found is returned.

DiameterLength

See ?getClusterStats().

Assortativity

See ?getClusterStats().

GlobalTransitivity

See ?getClusterStats().

EdgeDensity

See ?getClusterStats().

DegreeCentralityIndex

See ?getClusterStats().

ClosenessCentralityIndex

See ?getClusterStats().

EigenCentralityIndex

See ?getClusterStats().

EigenCentralityEigenvalue

See ?getClusterStats().

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

https://www.frontiersin.org/articles/10.3389/fimmu.2023.1181825

https://github.com/mlizhangx/Network-Analysis-for-Repertoire-Sequencing-

Author

Brian Neal (Brian.Neal@ucsf.edu)

See also

Examples

if (FALSE) {
## Generate some toy data for demonstration

# Use temp dir
data_dir <- tempdir()

# Directory to store input files
dir_input_samples <- file.path(data_dir, "input_samples")
dir.create(dir_input_samples, showWarnings = FALSE)

samples <- 30
affixes <- c("AAAA", "AASA", "AACA", "AAQA", "AAQ", "AAA", "AASAA", "AAAAA")
affix_probs_g0 <- rep(1 / length(affixes),
                      times = length(affixes) * samples / 2)
affix_probs_g1 <- rep(c(1, 5, 1, 1, 1, 1, 5, 1), times = samples / 2)
affix_probs <- matrix(c(affix_probs_g0, affix_probs_g1),
                      nrow = samples, byrow = TRUE)
new_probs_g0 <- rep(c(1/2, 1/6, 1/6, 1/6), times = samples / 2)
new_probs_g1 <- rep(c(1/3, 1/6, 1/6, 1/3), times = samples / 2)
new_probs <- matrix(c(new_probs_g0, new_probs_g1),
                    nrow = samples, byrow = TRUE)
simulateToyData(
  samples = samples,
  sample_size = 30,
  prefix_length = 1,
  prefix_chars = c("A", "C"),
  prefix_probs = cbind(rep(1, samples), rep(0, samples)),
  affixes = affixes,
  affix_probs = affix_probs,
  num_edits = 4,
  edit_pos_probs = function(seq_length) {
    dnorm(seq(-4, 4, length.out = seq_length))
  },
  edit_ops = c("insertion", "deletion", "transmutation"),
  edit_probs = c(5, 1, 4),
  new_chars = c("A", "S", "C", "Q"),
  new_probs = new_probs,
  output_dir = dir_input_samples,
  no_return = TRUE
)

## 1. Find Public Clusters in Each Sample

# input files for step 1 (one per sample)
input_files <- file.path(dir_input_samples, paste0("Sample", 1:samples, ".rds"))
head(input_files)

# Search across samples for public clusters
dir_filtered_samples <- file.path(data_dir, "filtered_samples")
findPublicClusters(
  file_list = input_files, input_type = "rds",
  sample_ids = paste0("Sample", 1:samples),
  seq_col = "CloneSeq", count_col = "CloneCount",
  min_seq_length = NULL, drop_matches = NULL,
  output_dir = dir_filtered_samples)

## 2. Build Public Cluster Network by Representative Sequence

# Cluster-level meta data for each sample's public clusters
dir_filtered_samples_cluster <- file.path(dir_filtered_samples, "cluster_meta_data")
files_filtered_samples_cluster <- list.files(dir_filtered_samples_cluster,
                                             full.names = TRUE)
head(files_filtered_samples_cluster)

dir_out <- file.path(data_dir, "public_clusters")

buildPublicClusterNetworkByRepresentative(
  file_list = files_filtered_samples_cluster,
  color_nodes_by = "ClusterIDPublic", output_dir = dir_out,
  size_nodes_by = 1)
}