Immune Repertoire Network By Sequence Similarity

Builds the network graph for an immune repertoire based on sequence similarity, computes specified network properties and generates customized visualizations.

Usage

buildRepSeqNetwork(

  ## Input ##
  data, seq_col, count_col = NULL,
  subset_cols = NULL,
  min_seq_length = 3, drop_matches = NULL,

  ## Network ##
  dist_type = "hamming", dist_cutoff = 1,
  drop_isolated_nodes = TRUE,
  node_stats = FALSE,
  stats_to_include = chooseNodeStats(),
  cluster_stats = FALSE,
  cluster_fun = cluster_fast_greedy,

  ## Visualization ##
  plots = TRUE,
  print_plots = TRUE,
  plot_title = "auto",
  plot_subtitle = "auto",
  color_nodes_by = "auto",
  ...,

  ## Output ##
  output_dir = getwd(),
  output_type = "individual",
  output_name = "MyRepSeqNetwork",
  pdf_width = 12, pdf_height = 10

)

Arguments

data

A data frame containing the immune repertoire sequencing data, with variables indexed by column and observations (e.g., clones or cells) indexed by row.

seq_col

The column name or number of data containing the receptor sequences to be used as the basis of similarity between rows. Also accepts a vector of length 2 specifying distinct sequence columns (e.g., alpha, beta chains); then two rows are similar only if both types of sequences are similar.

count_col

Optional column name or number of data containing a measure of abundance, e.g., clone count. Passed to getClusterStats; only relevant if cluster_stats = TRUE.

subset_cols

Optional vector of column names or numbers of data; if supplied, only these plus other relevant columns are included in the output. If NULL (the default), all columns of data are included.

min_seq_length

A numeric scalar, or NULL. Observations whose receptor sequences have fewer than min_seq_length characters are removed prior to network analysis.

drop_matches

An optional regular expression or character string. Observations whose receptor sequences return a match are removed prior to network analysis.

dist_type

The type of function to use as a measure of similarity between two receptor sequences. Valid options are "hamming" (the default), "levenshtein" and "euclidean_on_atchley" (only applicable to TCR CDR3 amino acid sequences).

dist_cutoff

A nonnegative scalar specifying the maximum distance threshold for similarity between receptor sequences.

drop_isolated_nodes

A logical scalar; should observations whose receptor sequences are not similar to any other sequences be dropped from the network?

node_stats

A logical scalar; should node-level network properties be computed?

stats_to_include

A logical vector returned by chooseNodeStats(), specifying the node-level properties to include. Also accepts the values "all" and "cluster_id_only". Only relevant if node_stats = TRUE.

cluster_stats

A logical scalar; should cluster-level network properties be computed?

cluster_fun

Passed to addClusterMembership. Controls the algorithm used for cluster identification.

plots

A logical scalar; should plot(s) of the network graph be generated?

print_plots

A logical scalar; should visualizations be printed in the R plotting window?

plot_title

A character string (or NULL value) to be used as the title in visualizations. The default value "auto" generates the title dynamically.

plot_subtitle

A character string (or NULL value) to be used as the subtitle in visualizations. The default value "auto" generates the title dynamically.

color_nodes_by

Optional. Column name or number of data used to color network nodes in visualizations. If a vector of values is supplied, one plot will be generated for each value. The value "auto" attempts to use one of several variables (e.g., cluster ID) to color the nodes, depending on what is available.

...

Other arguments to generateNetworkGraphPlots.

output_dir

The output directory; if NULL, output will not be written to file.

output_type

A character string specifying the file format to use when writing output to file. Default "individual" saves each item as a separate, uncompressed file, with data frames saved in csv format. "rda" and "rds" save the output list as a rda and rds file, respectively. For all output types, any plots generated will also be saved in a pdf file.

output_name

A character string to be used as a common filename prefix for any files saved.

pdf_width

Passed to the width argument of grDevices::pdf().

pdf_height

Passed to the height argument of grDevices::pdf().

Details

To build the immune repertoire network, each TCR/BCR clone (bulk data) or cell (single-cell data) is modeled as a network node (corresponding to a single row of the input data). Two nodes are considered adjacent (share an edge) if their receptor sequences are sufficiently similar.

Both nucleotide and amino acid sequences are supported. Sequence similarity is based on either the Hamming distance or Levenshtein (edit) distance. For TCR CDR3 amino acid sequences, an alternative measure of similarity is also available via the argument dist_type = "euclidean_on_atchley"; this represents sequences as 30-dimensional numeric vectors according to the Atchley factors of their amino acids (which encode biological properties) using a trained encoder. Similarity is then based on the Euclidean distance between these representations. Selecting a distance threshold is challenging in this setting, however, as particular values lack the clear interpretations enjoyed by the Hamming and Levenshtein distances.

The graph adjacency matrix for the network is computed and used to generate a igraph network object. Network properties are computed largely through the use of igraph functions. The visualizations are generated with ggraph.

Value

If the constructed network contains no edges, the function will return NULL with a warning. Otherwise, the function invisibly returns a list containing some or all of the following items:

igraph

An igraph object containing the edge list for the network.

adjacency_matrix

The network graph adjacency matrix, stored as a sparse matrix of class dgCMatrix from the Matrix package.

node_data

A data frame containing the node-level meta-data for the network. This data frame contains all column names of data unless subset_cols is non-null, in which case only the column names of data specified by subset_cols, along with other relevant ones (e.g., seq_col), will be present in the returned data frame. The data frame will additionally contain columns for any node-level network properties computed (as specified through the node_stats and stats_to_include arguments). The data frame will contain one row for each node that remains in the network after all filtering has occurred, including any filtering based on drop_isolated_nodes. The row names of the original input data will be preserved.

cluster_data

A data frame containing the cluster-level properties, with one row per cluster. Only included if cluster_stats = TRUE.

plots

A list containing the graph plot(s) as ggraph objects. Only included if plots = TRUE.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

https://www.frontiersin.org/articles/10.3389/fimmu.2023.1181825

https://github.com/mlizhangx/Network-Analysis-for-Repertoire-Sequencing-

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

if (FALSE) {

# Simulate some toy data for demonstration
toy_data <- simulateToyData()

network <- buildRepSeqNetwork(
  toy_data, seq_col = "CloneSeq",
  node_stats = TRUE, cluster_stats = TRUE,
  color_nodes_by = "transitivity", color_scheme = "plasma-1",
  size_nodes_by = "degree", node_size_limits = c(0.5, 1.5),
  output_dir = NULL)
}