Load Multiple Data Files and Combine

Given multiple data frames stored in separate files, loadDataFromFileList loads and combines them into a single data frame.

combineSamples is similar to loadDataFromFileList, but allows the data frames to be filtered and subsetted before being combined, and can automatically add sample-level variables such as sample ID.

Usage

loadDataFromFileList(
    file_list, input_type,
    data_symbols = NULL,
    header = TRUE, sep = "")

combineSamples(
    file_list, input_type,
    data_symbols = NULL,
    header = TRUE, sep = "",
    seq_col,
    min_seq_length = NULL,
    drop_matches = NULL,
    subset_cols = NULL,
    sample_ids = NULL,
    subject_ids = NULL,
    group_ids = NULL)

Arguments

file_list

A character vector of file paths containing the data frames (one file per data frame).

input_type

A character string specifying the input file format; determines the function used to load each file. See details section below.

data_symbols

If input_type = "rda", a character vector specifying the name of each data frame (this can be of length 1 if all data frames have the same name).

header

Passed to read.table or read.csv if applicable.

sep

Passed to read.table or read.csv if applicable.

seq_col

Passed to filterInputData for each sample.

min_seq_length

Passed to filterInputData for each sample.

drop_matches

Passed to filterInputData for each sample.

subset_cols

Passed to filterInputData for each sample.

sample_ids

An optional character or numeric vector of sample IDs, whose length matches that of file_list.

subject_ids

An optional character or numeric vector of subject IDs, whose length matches that of file_list.

group_ids

An optional character or numeric vector of group IDs, whose length matches that of file_list.

Details

Valid options for input_type (and the corresponding function used to load each file) include "rds" (readRDS), "rda" (load), "csv" (read.csv) and "table" (read.table).

If input_type = "rda", the data_symbols argument specifies the name of each data frame in R. For example, given three R data frames named sample1, sample2 and sample3, with each data frame saved to a separate rda file using save()), we would pass c("sample1", "sample2", "sample3") to data_symbols.

For combineSamples, for each of sample_ids, subject_ids and group_ids that is non-null, a corresponding variable will be added to the combined data frame; these will be named SampleID, SubjectID and GroupID.

Value

A data frame containing the combined data rows from all files.

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

## Generate some toy data for demonstration

# Use temp dir
data_dir <- tempdir()

# Directory to store input files
dir_samples <- file.path(data_dir, "samples")
dir.create(dir_samples, showWarnings = FALSE)
simulateToyData(sample_size = 5,
                output_dir = dir_samples,
                no_return = TRUE)
#> [1] TRUE

# Load data frames and combine
data <- loadDataFromFileList(
  file_list = list.files(dir_samples, full.names = TRUE),
  input_type = "rds")

## Same as above, but filter out seqs shorter than 3 characters,
## drop the count/freq/sampleID columns & add new sample ID column
data2 <- combineSamples(
  file_list = list.files(dir_samples, full.names = TRUE),
  input_type = "rds",
  seq_col = "CloneSeq",
  min_seq_length = 3,
  subset_cols = "CloneSeq",
  sample_ids = c("id01", "id02"))
#> >>> Loading and compiling data from all samples:
#> Loading sample 1: Input data contains 5 rows.
#> Removing sequences with length fewer than 3 characters... Done. 5 rows remaining.
#> Loading sample 2: Input data contains 5 rows.
#> Removing sequences with length fewer than 3 characters... Done. 5 rows remaining.
#> All samples loaded.