Encode TCR Sequences Numerically Based on Atchley Factor Representations

Given a list of T-cell receptor CDR3 amino acid sequences, encodes each sequence as a 30-dimensional numeric vector based on the Atchley factor representations of its amino acids.

Usage

encodeTCRSeqsByAtchleyFactor(
    cdr3_AA,
    contig_ids = seq_along(cdr3_AA)
)

Arguments

cdr3_AA

A character vector containing the TCR CDR3 amino acid sequences.

contig_ids

A numeric vector of the same length as cdr3_AA, used to uniquely identify its elements. By default, this is simply the sequence of integers from 1 to length(cdr3_AA).

Details

The encoding is performed using the software TESSA. For each TCR CDR3 amino acid sequence, each of its amino acids is represented by its five-dimensional vector of Atchley factors, which numerically encode its biological characteristics. The TCR CDR3 amino acid sequence is then encoded as a 30-dimensional numeric vector based on the Atchley factor representations of its amino acids using a deep learning algorithm.

Value

A matrix with length(cdr3_AA) rows and 31 columns. Each row corresponds to a TCR sequence; the first column is contig_ids, while the remaining columns contain the 30-dimensional numerical encoding for each TCR sequence.

Note

The deep learning algorithm used by TESSA was trained specifically on TCR CDR3 amino acid sequences and is not appropriate for use with other amino acid sequences.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

Zhang Z, Xiong D, Wang X, Liu H, Wang T. (2021). Mapping the functional landscape of T cell receptor repertoires by single-T cell transcriptomics. Nat Methods, 18(1):92–9. https://doi.org/10.1038/s41592-020-01020-3

Atchley WR, Zhao J, Fernandes AD, Druke T. (2005). Solving the protein sequence metric problem. Proc Natl Acad Sci U.S.A. 102(18):6395–400. https://doi.org/10.1073/pnas.0408677102

Webpage for the NAIR package

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

if (FALSE) {
cdr3 <- c("CASSEAQGSGSTDTQYF",
          "CATTEGSNTGELFF",
          "CASSIGDNEQFF",
          "CATSRDPDRGQSDTQYF",
          "CASSPTGLSGNTIYF",
          "CASSEEAGKDTQYF",
          "CASSGGADTQYF",
          "CASSLGLATDTQYF",
          "CASSEKEEVGELFF",
          "CASSSRTSGGAGELFF")

encodeTCRSeqsByAtchleyFactor(cdr3)
}