Filter RepSeq Data Rows and/or Subset Columns

Filter data rows by sequence length and/or sequence content; subset data columns using a numeric or character vector of column references.

Usage

filterInputData(data, seq_col, min_seq_length = NULL,
                drop_matches = NULL,
                subset_cols = NULL,
                count_col = NULL)

Arguments

data

A data frame containing the immune repertoire sequencing data, with variables indexed by column and observations (e.g., clones or cells) indexed by row.

seq_col

The column name of data containing the receptor sequences for use in any filtering based on length or content.

min_seq_length

A numeric scalar, or NULL. Observations whose receptor sequences have fewer than min_seq_length characters are removed prior to network analysis.

drop_matches

An optional regular expression or character string. Observations whose receptor sequences return a match are removed prior to network analysis.

subset_cols

Optional vector of column names or numbers of data; if supplied, only these (in addition to the column referenced by seq_col) are included in the output. If NULL (the default), all columns of data are included.

count_col

Optional column name or number of data; if supplied, the specified column of the data will be coerced to numeric and data rows containing NA/NaNs in the count column will be dropped.

Details

The sequence column(s) of the data are coerced to character type and rows containing NA values in the sequence column(s) are dropped.

Value

A data frame containing the filtered rows and specified columns of data.

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

## Generate some toy data for demonstration
raw_data <- simulateToyData()

# Filter data
filtered_data <-
  filterInputData(raw_data, seq_col = "CloneSeq",
                  min_seq_length = 13,
                  drop_matches = "[*|_]",
                  subset_cols = c("CloneSeq", "CloneFrequency", "SampleID"))
#> Input data contains 200 rows.
#> Removing sequences with length fewer than 13 characters... Done. 136 rows remaining.
#> Removing sequences containing matches to the expression '[*|_]'... Done. 136 rows remaining.