Analyze TCR Samples Using Numerical Encoder and K-Means Clustering

Given samples of TCR RepSeq data, convert the CDR3 amino acid sequences into numeric vectors using deep learning, perform \(K\)-means clustering, profile how each sample's unique TCR sequences are distributed among the clusters, and compare these profiles across samples.

Usage

kmeansAtchley(
  data,
  amino_col = "AminoAcidSeq",
  sample_col = "SampleID",
  group_col, k = 100,
  pdf_width = 15,
  pdf_height = 15,
  margin_cluster_heatmap = 25,
  margin_corr_heatmap = 15,
  use_viridis = FALSE,
  output_dir = getwd(),
  file_cluster_heatmap =
    "atchley_kmeans_cluster_relative_size_profiles_by_sample.pdf",
  file_corr_heatmap =
    "atchley_kmeans_corr_in_cluster_size_profile_between_samples.pdf",
  return_output = FALSE)

Arguments

data

A data frame containing the combined immune repertoire sequencing data for all samples, with variables indexed by column and observations indexed by row.

amino_col

The column of data containing the CDR3 amino acid sequences.

sample_col

The column of data containing the sample IDs.

group_col

The column of data containing the group IDs.

k

The number of clusters for \(K\)-means clustering.

pdf_width

Passed to the width argument of grDevices::pdf().

pdf_height

Passed to the height argument of grDevices::pdf().

margin_cluster_heatmap

The margin width of column names and row names in the heatmap for each sample's cluster profile.

margin_corr_heatmap

The margin width of column names and row names in the heatmap for the correlation in cluster profiles between samples.

use_viridis

A logical scalar indicating whether to use the cividis color scale from the viridis package instead of the RdBu color scale from RColorBrewer. This yields more readable results when printing in black-and-white, and is more robust against color vision deficiency.

output_dir

The directory in which to save pdf files of the two heatmaps produced.

file_cluster_heatmap

The pdf filename for the heatmap for each sample's cluster profile.

file_corr_heatmap

The pdf filename for the heatmap for correlation in cluster profiles between samples.

return_output

A logical scalar indicating whether to return the numeric vectors encoding the TCR sequences, the K-means cluster membership table, and each sample's profile of TCR representation across clusters.

Details

Each unique TCR sequence is encoded as a 30-dimensional numeric vector using encodeTCRSeqsByAtchleyFactor(), and \(K\)-means clustering is performed on the encoded TCR sequences from all samples, resulting in k clusters.

For each sample, its TCR representation profile across the k clusters is computed as a length-k vector whose \(i\)th element is the fraction of the sample's unique TCR sequences that belong to the \(i\)th cluster.

Two heatmaps are generated and saved to file: the first shows the TCR representation profile values across samples and clusters; the second shows the correlation in TCR represenation profiles between pairs of samples.

Value

If return_output = TRUE, then a list containing the following items:

kmeans_cluster_ids

A data frame with two variables, cdr3 and kmeanClusterID, containing the unique TCR sequences and the cluster to which each belongs.

encoded_values

A matrix returned by encodeTCRSeqsByAtchleyFactor containing the numerically encoded values of the TCR sequences.

cluster_TCR_profiles

A matrix with k rows and one column per sample; each column is the TCR representation profile for a sample, with each value recording the fraction of that sample's unique TCR sequences belonging to the cluster for the corresponding row (thus each column sums to 1).

References

The BriseisEncoder is authored by Ze Zhang (Ze.Zhang@UTsouthwestern.edu).

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

if (FALSE) {
toy_data <- simulateToyData(
  samples = 20,
  sample_size = 50,
  prefix_length = 0,
  prefix_chars = "",
  prefix_probs = matrix(1, nrow = 20),
  affixes = c("CASSLGYEQYF", "CASSLGETQYF",
              "CASSLGTDTQYF", "CASSLGTEAFF",
              "CASSLGGTEAFF", "CAGLGGRDQETQYF",
              "CASSQETQYF", "CASSLTDTQYF",
              "CANYGYTF", "CANTGELFF",
              "CSANYGYTF"),
  affix_probs = matrix(1, ncol = 11, nrow = 20),
)
toy_data$GroupID <- rep(c("G1", "G2"), each = 500)

atchley <- kmeansAtchley(
  data = toy_data,
  k = 3,
  amino_col = "CloneSeq",
  sample_col = "SampleID",
  group_col = "GroupID",
  output_dir = tempdir(),
  return_output = TRUE)
}