Save List of Network Objects

Given a list of network objects such as that returned by buildRepSeqNetwork or generateNetworkObjects, saves its contents according to the specified file format scheme.

Usage

saveNetwork(net, output_dir = getwd(),
            output_type = "individual",
            output_filename = "MyRepSeqNetwork",
            pdf_width = 12, pdf_height = 10)

Arguments

net

A list conforming to the format returned by buildRepSeqNetwork or generateNetworkObjects. See details.

output_dir

A file path specifying the directory in which to save the file(s).

output_type

A character string specifying the file format to use when writing output to file. Valid options are "individual", "rds" and "rda".

output_filename

A character string to be used as a common filename prefix for any files saved.

pdf_width

Passed to the width argument of the pdf function.

pdf_height

Passed to the height argument of the pdf function.

Details

The list provided to the net argument must contain at least the components igraph (an igraph object), adjacency_matrix (a matrix or dgCMatrix) and node_data (a data frame); the components cluster_data (a data frame) and plots (a list of ggraph objects) are optional, and will also be saved if present.

The saves each item as a separate, uncompressed file, with data frames saved in csv format. Setting output_type to "rda" or "rds" saves the output list as a rda or rds file, respectively. For all output types, any plots present will also be saved in a pdf file.

Under the default file format scheme output_type = "individual", the igraph is saved in edgelist format as a .txt file; the adjacency matrix is saved either in .csv format (for dense matrices) or .mtx format (for sparse matrices); and plots are saved only in a .pdf file, meaning that any plots will need to be generated again in order to be modified. If output_type is "rds" or "rda", the entire list net will be saved (the R symbol for the object will be net if saving as .rda).

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

if (FALSE) {
# Simulate some toy data for demonstration
toy_data <- simulateToyData()

# Generate network for data
net <- generateNetworkObjects(toy_data, "CloneSeq")

# Save network objects
saveNetwork(net, output_dir = tempdir())
}