Compute Network Adjacency Matrix for Immune Repertoire

Given a list of receptor sequences, computes the adjacency matrix for the network graph based on sequence similarity.

Usage

sparseAdjacencyMatFromSeqs(
    seqs,
    dist_type = "hamming",
    max_dist = 1,
    drop_isolated_nodes = TRUE)

Arguments

seqs

A character vector containing the receptor sequences.

dist_type

A character string specifying the distance function used to measure similarity between sequences. Valid options are "hamming" and "levenshtein".

max_dist

A nonnegative integer. Pairs of sequences whose distance is less than or equal to this value will be adjacent (connected by an edge) in the network graph.

drop_isolated_nodes

A logical scalar; if TRUE, sequences that are not adjacent to any other sequence will be excluded from the network.

Details

The adjacency matrix of a graph with \(n\) nodes/vertices is the symmetric \(n \times n\) matrix for which the \((i,j)\)th entry is equal to 1 if nodes \(i\) and \(j\) are adjacent (joined by an edge) in the network graph and 0 otherwise.

Each node in the network graph represents a receptor sequence from the seqs vector. Two nodes are adjacent if the distance between their sequences, as measured by dist_type, is less than or equal to max_dist.

Value

A sparse matrix of class dgCMatrix from the Matrix package.

If drop_isolated_nodes = TRUE, the row and column names of the matrix indicate which receptor sequences in the seqs vector correspond to each row and column. The row and column names can be accessed using dimnames(). This returns a list containing two character vectors, one for the row names and one for the column names. The name of the \(i\)th row is the numerical index of the seqs vector corresponding to the \(i\)th row and \(i\)th column. The name of the \(j\)th column is the receptor sequence corresponding to the \(j\)th row and \(j\)th column.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

https://www.frontiersin.org/articles/10.3389/fimmu.2023.1181825

https://github.com/mlizhangx/Network-Analysis-for-Repertoire-Sequencing-

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

sparseAdjacencyMatFromSeqs(
    c("fee", "fie", "foe", "fum", "foo")
)
#> Computing network edges based on a max hamming distance of 1... Done.
#> Network contains 4 nodes (after removing isolated nodes).
#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>   fee fie foe foo
#> 1   1   1   1   .
#> 2   1   1   1   .
#> 3   1   1   1   1
#> 5   .   .   1   1