Perform Cluster Analysis on the Nodes of a Network Graph

Given the igraph and node-level metadata for a network, partitions the graph into densely-connected subgraphs (clusters) using the specified clustering algorithm, and adds a cluster membership ID variable to the node-level metadata.

Usage

addClusterMembership(
  data,
  net,
  fun = cluster_fast_greedy
)

Arguments

data

A data frame containing the node-level metadata for the network, such as the data frame named node_data contained in the output of buildRepSeqNetwork or generateNetworkObjects.

net

The igraph object for the network, such as the element named igraph contained in the output of buildRepSeqNetwork or generateNetworkObjects.

fun

The function to use for the clustering algorithm. See clustering_algorithms.

Details

Cluster analysis involves using a community-finding algorithm to partition the network graph into clusters (densely-connected subgraphs). These clusters represent groups of clones/cells with similar receptor sequences.

Each cluster is assigned a numeric cluster ID, and the cluster membership ID of each node is recorded. The cluster membership ID is the cluster ID of the cluster to which the node belongs.

Refer to the help topic clustering_algorithms for information on the available clustering algorithms that can be used to perform cluster analysis.

Value

A copy of data with an additional column named cluster_id that contains the cluster membership ID for the network node corresponding to each row.

See also

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. doi: 10.3389/fimmu.2023.1181825

Webpage for the NAIR package

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

if (FALSE) {
toy_data <- simulateToyData()

# Generate network
net <-
  generateNetworkObjects(
    toy_data,
    "CloneSeq"
  )

# Perform cluster analysis,
# add cluster membership ID
data_w_clusterID <-
  addClusterMembership(
    net$node_data,
    net$igraph
  )
}