Build Global Network of Public TCR/BCR Clusters

Part of the workflow Searching for Public TCR/BCR Clusters. Intended for use following findPublicClusters.

Given node-level metadata for each sample's filtered clusters, combines the data into a global network and performs network analysis and cluster analysis.

Usage

buildPublicClusterNetwork(

  ## Input ##
  file_list =
    list.files(file.path(getwd(), "public_clusters", "node_meta_data")),
  input_type = "rds",
  data_symbols = "ndat",
  header = TRUE, sep = "",
  seq_col,

  ## Network Settings ##
  drop_isolated_nodes = FALSE,
  node_stats = NULL, stats_to_include = NULL, cluster_stats = NULL,

  ## Visualization ##
  color_nodes_by = "SampleID",
  color_scheme = "turbo",
  plot_title = "Global Network of Public Clusters",

  ## Output ##
  output_dir =
    file.path(getwd(), "public_clusters"),
  output_name = "PublicClusterNetwork",

  ...

)

Arguments

file_list

A vector of file paths where each file contains the node-level metadata for one sample's filtered clusters. Passed to loadDataFromFileList.

input_type

A character string specifying the file format of the input files. Options are "csv", "rds" and "rda". Passed to loadDataFromFileList.

data_symbols

Used when input_type = "rda". Specifies the name of the data frame within each Rdata file. Passed to loadDataFromFileList.

header

Used for certain values of input_type. Passed to loadDataFromFileList.

sep

Used for certain values of input_type. Passed to loadDataFromFileList.

seq_col

Specifies the column in the node-level metadata that contains the TCR/BCR sequences. Accepts either the column name or column index.

drop_isolated_nodes

Passed to buildRepSeqNetwork when constructing the global network.

node_stats

Deprecated. Does nothing since all network properties are automatically computed.

stats_to_include

Deprecated. Does nothing since all network properties are automatically computed.

cluster_stats

Deprecated. Does nothing since all network properties are automatically computed.

color_nodes_by

Passed to buildRepSeqNetwork when constructing the global network.

color_scheme

Passed to buildRepSeqNetwork when constructing the global network.

plot_title

Passed to buildRepSeqNetwork when constructing the global network.

output_dir

Passed to buildRepSeqNetwork when constructing the global network.

output_name

Passed to buildRepSeqNetwork when constructing the global network.

...

Other arguments to buildRepSeqNetwork when constructing the global network.

Details

The node-level metadata for the filtered clusters from all samples is combined in order to construct a single global network. Cluster analysis is used to partition the global network into clusters, which are considered as the public TCR/BCR clusters. Network properties for the nodes and clusters are computed and returned as metadata. A plot of the global network graph is produced, with the nodes colored according to sample ID.

The computed node-level network properties are renamed to reflect their correspondence to the global network. This is done to distinguish them from the network properties that correspond to the sample-level networks. The names are ClusterIDPublic, PublicNetworkDegree, PublicTransitivity, PublicCloseness, PublicCentralityByCloseness, PublicEigenCentrality, PublicCentralityByEigen, PublicBetweenness, PublicCentralityByBetweenness, PublicAuthorityScore, PublicCoreness, PublicPageRank.

See the Searching for Public TCR/BCR Clusters package vignette for more details. The vignette can be accessed offline from using vignette("public_clusters").

Value

A list of network objects as returned by buildRepSeqNetwork.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

Webpage for the NAIR package

Searching for Public TCR/BCR Clusters vignette

Author

Brian Neal (Brian.Neal@ucsf.edu)

See also