Build Global Network of Public TCR/BCR Clusters Using Representative Clones

Alternative step in the workflow Searching for Public TCR/BCR Clusters. Intended for use following findPublicClusters in cases where buildPublicClusterNetwork cannot be practically used due to the size of the full global network.

Given cluster-level metadata for each sample's filtered clusters, selects a representative TCR/BCR from each cluster, combines the representatives into a global network and performs network analysis and cluster analysis.

Usage

buildPublicClusterNetworkByRepresentative(

  ## Input ##
  file_list,
  input_type = "rds",
  data_symbols = "cdat",
  header = TRUE, sep = "",
  seq_col = "seq_w_max_count",
  count_col = "agg_count",

  ## Network Settings ##
  dist_type = "hamming",
  dist_cutoff = 1,
  cluster_fun = cluster_fast_greedy,

  ## Visualization ##
  plots = TRUE,
  print_plots = TRUE,
  plot_title = "auto",
  plot_subtitle = "auto",
  color_nodes_by = "SampleID",
  color_scheme = "turbo",
  ...,

  ## Output ##
  output_dir =
    file.path(getwd(), "public_clusters"),
  output_type = "rda",
  output_name = "PubClustByRepresentative",
  pdf_width = 12,
  pdf_height = 10

)

Arguments

file_list

A vector of file paths where each file contains the cluster-level metadata for one sample's filtered clusters. Passed to loadDataFromFileList.

input_type

A character string specifying the file format of the input files. Options are "csv", "rds" and "rda". Passed to loadDataFromFileList.

data_symbols

Used when input_type = "rda". Specifies the name of the data frame within each Rdata file. Passed to loadDataFromFileList.

header

Used for certain values of input_type. Passed to loadDataFromFileList.

sep

Used for certain values of input_type. Passed to loadDataFromFileList.

seq_col

Specifies the column in the cluster-level metadata that contains the representative TCR/BCR sequence for each cluster. Accepts either the column name or column index. By default, uses the sequence with the maximum clone count in each cluster.

count_col

Specifies the column in the cluster-level metadata that contains the aggregate clone count for each cluster. Accepts either the column name or column index.

dist_type

Passed to buildRepSeqNetwork when constructing the global network.

dist_cutoff

Passed to buildRepSeqNetwork when constructing the global network.

cluster_fun

Passed to buildRepSeqNetwork when performing cluster analysis on the global network.

plots

Logical. Should plots of the global network graph be produced?

print_plots

Logical. If plots of the global network graph are produced, should they be printed to the R plotting window?

plot_title

Passed to generateNetworkGraphPlots when producing plots of the global network graph.

plot_subtitle

Passed to generateNetworkGraphPlots when producing plots of the global network graph.

color_nodes_by

Passed to generateNetworkGraphPlots when producing plots of the global network graph.

color_scheme

Passed to generateNetworkGraphPlots when producing plots of the global network graph.

...

Other arguments to generateNetworkGraphPlots when producing plots of the global network graph.

output_dir

Passed to saveNetwork after constructing the global network.

output_type

Passed to saveNetwork after constructing the global network.

output_name

Passed to saveNetwork after constructing the global network.

pdf_width

Passed to saveNetwork after constructing the global network. Only applicable if plots = TRUE.

pdf_height

Passed to saveNetwork after constructing the global network. Only applicable if plots = TRUE.

Details

From each filtered cluster in each sample's network, a representative TCR/BCR is selected. By default, this is the sequence with the greatest clone count in each cluster. The representatives from all clusters and all samples are then used to construct a single global network. Cluster analysis is used to partition this global network into clusters. Network properties for the nodes and clusters are computed and returned as metadata. A plot of the global network graph is produced, with the nodes colored according to sample ID.

Within this network, clusters containing nodes from multiple samples can be considered as the skeletons of the complete public clusters. The filtered cluster data for each sample can then be subset to keep the sample-level clusters whose representative TCR/BCRs belong to the skeletons of the public clusters. After subsetting in this manner, buildPublicClusterNetwork can be used to construct the global network of complete public clusters.

Value

A list of network objects as returned by buildRepSeqNetwork. The data frame cluster_data includes the following variables that represent properties of the clusters in the global network of representative TCR/BCRs:

TotalSampleLevelNodes

For each representative TCR/BCR in the global cluster, we record the number of nodes in the sample-level cluster for which it is the representative TCR/BCR. We then sum these node counts across all the representative TCR/BCRs in the global cluster.

TotalCloneCount

For each representative TCR/BCR in the global cluster, we record the aggregate clone count from all nodes in the sample-level cluster for which it is the representative TCR/BCR. We then sum these aggregate clone counts across all the representative TCR/BCRs in the global cluster.

MeanOfMeanSeqLength

For each representative TCR/BCR in the global cluster, we record the mean sequence length over all clones (nodes) in the sample-level cluster for which it is the representative TCR/BCR. We then average these mean sequence lengths over all the representative TCR/BCRs in the global cluster.

MeanDegreeInPublicNet

For each representative TCR/BCR in the global cluster, we record the mean network degree over all nodes in the sample-level cluster for which it is the representative TCR/BCR. We then average these mean degree values over all the representative TCR/BCRs in the global cluster.

MaxDegreeInPublicNet

For each representative TCR/BCR in the global cluster, we record the maximum network degree across all nodes in the sample-level cluster for which it is the representative TCR/BCR. We then take the maximum of these maximum degree values over all the representative TCR/BCRs in the global cluster.

SeqWithMaxDegree

For each representative TCR/BCR in the global cluster, we record the maximum network degree across all nodes in the sample-level cluster for which it is the representative TCR/BCR. We then identify the representative TCR/BCR with the maximum value of these maximum degrees over all the representative TCR/BCRs in the global cluster. The TCR/BCR sequence of the identified representative TCR/BCR is recorded in this variable.

MaxCloneCount

For each representative TCR/BCR in the global cluster, we record the maximum clone count across all clones (nodes) in the sample-level cluster for which it is the representative TCR/BCR. We then take the maximum of these maximum clone counts over all the representative TCR/BCRs in the global cluster.

SampleWithMaxCloneCount

For each representative TCR/BCR in the global cluster, we record the maximum clone count across all clones (nodes) in the sample-level cluster for which it is the representative TCR/BCR. We then identify the representative TCR/BCR with the maximum value of these maximum clone counts over all the representative TCR/BCRs in the global cluster. The sample to which the identified representative TCR/BCR belongs is recorded in this variable.

SeqWithMaxCloneCount

For each representative TCR/BCR in the global cluster, we record the maximum clone count across all clones (nodes) in the sample-level cluster for which it is the representative TCR/BCR. We then identify the representative TCR/BCR with the maximum value of these maximum clone counts over all the representative TCR/BCRs in the global cluster. The TCR/BCR sequence of the identified representative TCR/BCR is recorded in this variable.

MaxAggCloneCount

For each representative TCR/BCR in the global cluster, we record the aggregate clone count across all clones (nodes) in the sample-level cluster for which it is the representative TCR/BCR. We then take the maximum of these aggregate clone counts over all the representative TCR/BCRs in the global cluster.

SampleWithMaxAggCloneCount

For each representative TCR/BCR in the global cluster, we record the aggregate clone count across all clones (nodes) in the sample-level cluster for which it is the representative TCR/BCR. We then identify the representative TCR/BCR with the maximum value of these aggregate clone counts over all the representative TCR/BCRs in the global cluster. The sample to which the identified representative TCR/BCR belongs is recorded in this variable.

SeqWithMaxAggCloneCount

For each representative TCR/BCR in the global cluster, we record the aggregate clone count across all clones (nodes) in the sample-level cluster for which it is the representative TCR/BCR. We then identify the representative TCR/BCR with the maximum value of these aggregate clone counts over all the representative TCR/BCRs in the global cluster. The TCR/BCR sequence of the identified representative TCR/BCR is recorded in this variable.

DiameterLength

See getClusterStats. Based on edge connections between representative TCR/BCRs in the global cluster.

Assortativity

See getClusterStats. Based on edge connections between representative TCR/BCRs in the global cluster.

GlobalTransitivity

See getClusterStats. Based on edge connections between representative TCR/BCRs in the global cluster.

EdgeDensity

See getClusterStats. Based on edge connections between representative TCR/BCRs in the global cluster.

DegreeCentralityIndex

See getClusterStats. Based on edge connections between representative TCR/BCRs in the global cluster.

ClosenessCentralityIndex

See getClusterStats. Based on edge connections between representative TCR/BCRs in the global cluster.

EigenCentralityIndex

See getClusterStats. Based on edge connections between representative TCR/BCRs in the global cluster.

EigenCentralityEigenvalue

See getClusterStats. Based on edge connections between representative TCR/BCRs in the global cluster.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

Webpage for the NAIR package

Searching for Public TCR/BCR Clusters vignette

Author

Brian Neal (Brian.Neal@ucsf.edu)

See also

Examples

if (FALSE) {
## Generate some toy data for demonstration

# Use temp dir
data_dir <- tempdir()

# Directory to store input files
dir_input_samples <- file.path(data_dir, "input_samples")
dir.create(dir_input_samples, showWarnings = FALSE)

samples <- 30
sample_size <- 30 # (seqs per sample)

base_seqs <- c(
  "CASSIEGQLSTDTQYF", "CASSEEGQLSTDTQYF", "CASSSVETQYF",
  "CASSPEGQLSTDTQYF", "RASSLAGNTEAFF", "CASSHRGTDTQYF", "CASDAGVFQPQHF",
  "CASSLTSGYNEQFF", "CASSETGYNEQFF", "CASSLTGGNEQFF", "CASSYLTGYNEQFF",
  "CASSLTGNEQFF", "CASSLNGYNEQFF", "CASSFPWDGYGYTF", "CASTLARQGGELFF",
  "CASTLSRQGGELFF", "CSVELLPTGPLETSYNEQFF", "CSVELLPTGPSETSYNEQFF",
  "CVELLPTGPSETSYNEQFF", "CASLAGGRTQETQYF", "CASRLAGGRTQETQYF",
  "CASSLAGGRTETQYF", "CASSLAGGRTQETQYF", "CASSRLAGGRTQETQYF",
  "CASQYGGGNQPQHF", "CASSLGGGNQPQHF", "CASSNGGGNQPQHF", "CASSYGGGGNQPQHF",
  "CASSYGGGQPQHF", "CASSYKGGNQPQHF", "CASSYTGGGNQPQHF",
  "CAWSSQETQYF", "CASSSPETQYF", "CASSGAYEQYF", "CSVDLGKGNNEQFF")

# relative generation probabilities
pgen <- cbind(
  stats::toeplitz(0.6^(0:(sample_size - 1))),
  matrix(1, nrow = samples, ncol = length(base_seqs) - samples))

# Simulate the data
foo <- simulateToyData(
  samples = samples, sample_size = sample_size,
  prefix_length = 1, prefix_chars = c("", ""),
  prefix_probs = cbind(rep(1, samples), rep(0, samples)),
  affixes = base_seqs, affix_probs = pgen, num_edits = 0,
  output_dir = dir_input_samples, no_return = FALSE)

# View first few rows of data for sample 1
head(readRDS(file.path(dir_input_samples, "Sample1.rds")))

## 1. Find Public Clusters in Each Sample

# input files for step 1 (one per sample)
input_files <-
  file.path(dir_input_samples,
            paste0("Sample", 1:samples, ".rds"))
head(input_files)

# directory to save filtered sample data
dir_filtered_samples <- file.path(data_dir, "filtered_samples")

# Filter Clusters Within Each Sample
findPublicClusters(
  file_list = input_files, input_type = "rds",
  seq_col = "CloneSeq", count_col = "CloneCount",
  min_seq_length = NULL, drop_matches = NULL,
  top_n_clusters = 3, min_node_count = 5, min_clone_count = 15000,
  output_dir = dir_filtered_samples)

## 2. Build Public Cluster Network by Representative TCR/BCRs

# Cluster-level meta data for each sample's public clusters
dir_filtered_samples_cluster <-
  file.path(dir_filtered_samples, "cluster_meta_data")
files_filtered_samples_cluster <-
  list.files(dir_filtered_samples_cluster, full.names = TRUE)
head(files_filtered_samples_cluster)

dir_out <- file.path(data_dir, "public_clusters")

buildPublicClusterNetworkByRepresentative(
  file_list = files_filtered_samples_cluster,
  output_dir = dir_out,
  size_nodes_by = 1)
}