Load and Combine Data From Multiple Samples

Given multiple data frames stored in separate files, loadDataFromFileList loads and combines them into a single data frame.

combineSamples is similar to loadDataFromFileList, but allows the data frames to be filtered and subsetted before being combined, and can automatically add sample-level variables such as sample ID.

Usage

loadDataFromFileList(
    file_list,
    input_type,
    data_symbols = NULL,
    header = TRUE, sep = ""
)

combineSamples(
    file_list,
    input_type,
    data_symbols = NULL,
    header = TRUE, sep = "",
    seq_col,
    min_seq_length = NULL,
    drop_matches = NULL,
    subset_cols = NULL,
    sample_ids = NULL,
    subject_ids = NULL,
    group_ids = NULL
)

Arguments

file_list

A vector of file paths where each file contains the data for a sample, with observations indexed by row.

input_type

A character string specifying the file format of the sample data files. Options are "table", "txt", "tsv", "csv", "rds" and "rda".

data_symbols

Used when input_type = "rda". Specifies the name of each sample's data frame within its respective Rdata file. Accepts a character vector of the same length as file_list. Alternatively, a single character string can be used if all data frames have the same name. See details.

header

Passed to read.table or read.csv for applicable values of input_type.

sep

Passed to read.table or read.csv for applicable values of input_type.

seq_col

Passed to filterInputData for each sample.

min_seq_length

Passed to filterInputData for each sample.

drop_matches

Passed to filterInputData for each sample.

subset_cols

Passed to filterInputData for each sample.

sample_ids

A character or numeric vector of sample IDs, whose length matches that of file_list.

subject_ids

An optional character or numeric vector of subject IDs, whose length matches that of file_list. Used to assign a subject ID to each sample.

group_ids

A character or numeric vector of group IDs whose length matches that of file_list. Used to assign each sample to a group.

Details

Valid options for input_type (and the corresponding function used to load each file) include "rds" (readRDS), "rda" (load), "csv" (read.csv) and one of "table", "txt" or "tsv" (all of which are loaded using read.table).

If input_type = "rda", the data_symbols argument specifies the name (in R) of each data frame as it is saved within its respective file. For example, given three R data frames named sample1, sample2 and sample3 in the R environment, with each data frame saved to a separate rda file using save), we would pass c("sample1", "sample2", "sample3") to data_symbols, assuming this is the order in which their file paths appear in the file_list argument. If each data frame is named sample_data within its respective rda file, we would simply specify data_symbols = "sample_data".

When calling combineSamples, for each of sample_ids, subject_ids and group_ids that is non-null, a corresponding variable will be added to the combined data frame; these variables are named SampleID, SubjectID and GroupID.

Value

A data frame containing the combined data rows from all files.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. doi: 10.3389/fimmu.2023.1181825

Webpage for the NAIR package

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

simulateToyData(sample_size = 5,
                output_dir = tempdir(),
                no_return = TRUE)
#> [1] TRUE

# Load and combine data
all_samples <-
  loadDataFromFileList(
    file_list =
      file.path(tempdir(), paste0("Sample", 1:2, ".rds")),
    input_type = "rds",
  )

# Same as above, but also apply some filters,
# keep only the CloneSeq column and add SampleID column
all_samples <-
  combineSamples(
    file_list =
      file.path(tempdir(), paste0("Sample", 1:2, ".rds")),
    input_type = "rds",
    seq_col = "CloneSeq",
    min_seq_length = 13,
    drop_matches = "GGG",
    subset_cols = "CloneSeq",
    sample_ids = c("id01", "id02")
  )
#> >>> Loading and compiling data from all samples:
#> Loading sample 1: Input data contains 5 rows.
#> Removing sequences with length fewer than 13 characters... Done. 4 rows remaining.
#> Removing sequences containing matches to the expression 'GGG'... Done. 3 rows remaining.
#> Loading sample 2: Input data contains 5 rows.
#> Removing sequences with length fewer than 13 characters... Done. 2 rows remaining.
#> Removing sequences containing matches to the expression 'GGG'... Done. 2 rows remaining.
#> All samples loaded.

# clean up temp directory
file.remove(
  file.path(tempdir(), paste0("Sample", 1:2, ".rds"))
)
#> [1] TRUE TRUE