Identify TCR/BCR Sequences Associated With a Binary Variable

Part of the workflow Searching for Associated TCR/BCR Clusters.

Given multiple samples of bulk Adaptive Immune Receptor Repertoire Sequencing (AIRR-Seq) data and a binary variable of interest such as a disease condition, treatment or clinical outcome, identify receptor sequences that exhibit a statistically significant difference in frequency between the two levels of the binary variable.

findAssociatedSeqs is designed for use when each sample is stored in a separate file. findAssociatedSeqs2 is designed for use with a single data frame containing all samples.

Usage

findAssociatedSeqs(
  ## Input ##
  file_list,
  input_type,
  data_symbols = NULL,
  header = TRUE, sep = "",
  sample_ids = NULL,
  subject_ids = NULL,
  group_ids,
  groups = NULL,
  seq_col,
  freq_col = NULL,

  ## Search Criteria ##
  min_seq_length = 7,
  drop_matches = "[*|_]",
  min_sample_membership = 5,
  pval_cutoff = 0.05,

  ## Output ##
  outfile = "associated_seqs.csv")

findAssociatedSeqs2(
  ## Input ##
  data,
  seq_col,
  sample_col,
  subject_col = sample_col,
  group_col,
  groups = NULL,
  freq_col = NULL,

  ## Search Criteria ##
  min_seq_length = 7,
  drop_matches = "[*|_]",
  min_sample_membership = 5,
  pval_cutoff = 0.05,

  ## Ouptut ##
  outfile = "associated_seqs.csv")

Arguments

file_list

A vector of file paths where each file contains the data for a sample, with observations indexed by row. Passed to loadDataFromFileList.

input_type

A character string specifying the file format of the sample data files. Options are "table", "txt", "tsv", "csv", "rds" and "rda". Passed to loadDataFromFileList.

data_symbols

Used when input_type = "rda". Specifies the name of each sample's data frame within its respective Rdata file. Passed to loadDataFromFileList.

header

Used for certain values of input_type. Passed to loadDataFromFileList.

sep

Used for certain values of input_type. Passed to loadDataFromFileList.

sample_ids

Deprecated. Does nothing.

subject_ids

A character or numeric vector of subject IDs, whose length matches that of file_list. Only relevant when the binary variable of interest is subject-specific and multiple samples belong to the same subject.

group_ids

A character or numeric vector of group IDs containing exactly two unique values and with length matching that of file_list. The two groups correspond to the two values of the binary variable of interest.

groups

Deprecated. Does nothing.

seq_col

Specifies the column of each sample's data frame containing the TCR/BCR sequences. Accepts either the column name or column index.

freq_col

Optional. Specifies the column of each sample's data frame containing the clone frequency (i.e., clone count divided by the sum of the clone counts across all clones in the sample). Accepts either the column name or column index. If this argument is specified, the maximum clone frequency (across all samples) for each associated sequence will be included in the content of the label variable of the returned data frame.

min_seq_length

Controls the minimum TCR/BCR sequence length considered when searching for associated sequences. Passed to filterInputData.

drop_matches

Passed to filterInputData. Accepts a regular expression (see regex). Checks TCR/BCR sequences for a pattern match using grep. Those returning a match are excluded from consideration as associated sequences. It is recommended to filter out sequences containing special characters that are invalid for use in file names. By default, sequences containing either of the characters * or _ are dropped.

min_sample_membership

Controls the minimum number of samples in which a TCR/BCR sequence must be present in order to be considered when searching for associated sequences. Setting this value to NULL bypasses the check.

pval_cutoff

Controls the P-value cutoff below which an association is detected by Fisher's exact test (see details).

outfile

A file path for saving the output (using write.csv).

data

A data frame containing the combined AIRR-seq data for all samples.

sample_col

The column of data containing the sample IDs. Accepts the column name or column index.

subject_col

Optional. The column of data containing the subject IDs. Accepts the column name or column index. Only relevant when the binary variable of interest is subject-specific and multiple samples belong to the same subject.

group_col

The column of data containing the group IDs. Accepts the column name or column index. The groups correspond to the two values of the binary variable of interest. Thus there should be exactly two unique values in this column.

Details

The TCR/BCR sequences from all samples are first filtered according to minimum sequence length and sequence content based on the specified values in min_seq_length and drop_matches, respectively. The sequences are further filtered based on sample membership, removing sequences appearing in fewer than min_sample_membership samples.

For each remaining TCR/BCR sequence, a P-value is computed for Fisher's exact test of independence between the binary variable of interest and the presence of the sequence within a repertoire. The samples/subjects are divided into two groups based on the levels of the binary variable. If subject IDs are provided, then the test is based on the number of subjects in each group for whom the sequence appears in one of their samples. Without subject IDs, the test is based on the number of samples possessing the sequence in each group.

Fisher's exact test is performed using fisher.test. TCR/BCR sequences with a \(P\)-value below pval_cutoff are sorted by \(P\)-value and returned along with some additional information.

The returned ouput is intended for use with the findAssociatedClones function. See the Searching for Associated TCR/BCR Clusters package vignette for more details. The vignette can be accessed offline using vignette("associated_clusters").

Value

A data frame containing the TCR/BCR sequences found to be associated with the binary variable using Fisher's exact test (see details). Each row corresponds to a unique TCR/BCR sequence and includes the following variables:

ReceptorSeq

The unique receptor sequence

fisher_pvalue

The P-value on Fisher's exact test for independence between the receptor sequence and the binary variable of interest

shared_by_n_samples

The number of samples in which the sequence was observed

samples_g0

Of the samples in which the sequence was observed, the number of samples belonging to the first group

samples_g1

Of the samples in which the sequence was observed, the number of samples belonging to the second group

shared_by_n_subjects

The number of subjects in which the sequence was observed (only present if subject IDs are specified)

subjects_g0

Of the subjects in which the sequence was observed, the number of subjects belonging to the first group (only present if subject IDs are specified)

subjects_g1

Of the subjects in which the sequence was observed, the number of subjects belonging to the second group (only present if subject IDs are specified)

label

A character string summarizing the above information. Also includes the maximum in-sample clone frequency across all samples, if available

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

Webpage for the NAIR package

Searching for Associated TCR/BCR Clusters vignette

Author

Brian Neal (Brian.Neal@ucsf.edu)

See also

Examples

if (FALSE) {
## Generate some toy data for demonstration ##
# Use temp dir
data_dir <- tempdir()

# Directory to store input files
dir_input_samples <- file.path(data_dir, "input_samples")
dir.create(dir_input_samples, showWarnings = FALSE)

# Number of samples by control/treatment group
samples_c <- samples_t <- 15
samples <- samples_c + samples_t
sample_size <- 30 # (seqs per sample)

# sequences (first five are chosen to be associated with treatment)
base_seqs <- c("CASSGAYEQYF", "CSVDLGKGNNEQFF",
               "CASSIEGQLSTDTQYF",
               "CASSEEGQLSTDTQYF",
               "CASSPEGQLSTDTQYF",
               "RASSLAGNTEAFF", "CASSHRGTDTQYF", "CASDAGVFQPQHF")

# relative generation probabilities by control/treatment group
pgen_c <- matrix(rep(c(rep(1, 5), rep(30, 3)), times = samples_c),
                 nrow = samples_c, byrow = TRUE)
pgen_t <- matrix(rep(c(1, 1, rep(1/3, 3), rep(2, 3)), times = samples_t),
                 nrow = samples_t, byrow = TRUE)
pgen <- rbind(pgen_c, pgen_t)

# Simulate the data
simulateToyData(
  samples = samples, sample_size = sample_size,
  prefix_length = 1, prefix_chars = c("", ""),
  prefix_probs = cbind(rep(1, samples), rep(0, samples)),
  affixes = base_seqs, affix_probs = pgen, num_edits = 0,
  output_dir = dir_input_samples, no_return = TRUE)

# View first few rows of data for sample 1
head(readRDS(file.path(dir_input_samples, "Sample1.rds")))

## Step 1: Find Associated Sequences ##
# input files (one per sample)
input_files <- file.path(dir_input_samples, paste0("Sample", 1:samples, ".rds"))

# group labels for the samples
group_labels <- c(rep("reference", samples_c), rep("comparison", samples_t))

# search for associated sequences using Fisher's exact test
associated_seqs <- findAssociatedSeqs(
  file_list = input_files, input_type = "rds",
  group_ids = group_labels,
  seq_col = "CloneSeq",
  min_seq_length = NULL, drop_matches = NULL,
  min_sample_membership = 0,
  pval_cutoff = 0.1, outfile = NULL)

# view first few rows of output
head(associated_seqs)
}