Find Public Clusters Among RepSeq Samples

Part of the workflow Searching for Public TCR/BCR Clusters.

Given multiple samples of bulk Adaptive Immune Receptor Repertoire Sequencing (AIRR-Seq) data, construct the repertoire network for each sample. Within each sample's network, perform cluster analysis and filter the clusters based on node count and aggregate clone count.

Usage

findPublicClusters(

  ## Input ##
  file_list,
  input_type,
  data_symbols = NULL,
  header = TRUE, sep = "",
  sample_ids =
    paste0("Sample", 1:length(file_list)),
  seq_col,
  count_col = NULL,

  ## Search Criteria ##
  min_seq_length = 3,
  drop_matches = "[*|_]",
  top_n_clusters = 20,
  min_node_count = 10,
  min_clone_count = 100,

  ## Optional Visualization ##
  plots = FALSE,
  print_plots = FALSE,
  plot_title = "auto",
  color_nodes_by = "cluster_id",

  ## Output ##
  output_dir =
    file.path(getwd(), "public_clusters"),
  output_type = "rds",

  ## Optional Output ##
  output_dir_unfiltered = NULL,
  output_type_unfiltered = "rds",

  ...

)

Arguments

file_list

A vector of file paths where each file contains the data for a sample, with observations indexed by row. Passed to loadDataFromFileList.

input_type

A character string specifying the file format of the sample data files. Options are "table", "txt", "tsv", "csv", "rds" and "rda". Passed to loadDataFromFileList.

data_symbols

Used when input_type = "rda". Specifies the name of each sample's data frame within its respective Rdata file. Passed to loadDataFromFileList.

header

Used for certain values of input_type. Passed to loadDataFromFileList.

sep

Used for certain values of input_type. Passed to loadDataFromFileList.

sample_ids

A character or numeric vector of sample IDs, whose length matches that of file_list. The values should be valid for use as filenames and should avoid using the forward slash or backslash characters (/ or \).

seq_col

Specifies the column of each sample's data frame containing the TCR/BCR sequences. Accepts either the column name or column index.

count_col

Specifies the column of each sample's data frame containing the clone count (measure of clonal abundance). Accepts either the column name or column index.

min_seq_length

Passed to buildRepSeqNetwork when constructing the network for each sample.

drop_matches

Passed to buildRepSeqNetwork when constructing the network for each sample. Accepts a regular expression (see regex). Checks TCR/BCR sequences for a pattern match. Those returning a match are dropped. By default, sequences containing either of the characters * or _ are dropped.

top_n_clusters

The number of clusters from each sample to be automatically be included among the filtered clusters, based on greatest node count.

min_node_count

Clusters with at least this many nodes will be included among the filtered clusters.

min_clone_count

Clusters with an aggregate clone count of at least this value will be included among the filtered clusters.

plots

Passed to buildRepSeqNetwork when constructing the network for each sample.

print_plots

Passed to buildRepSeqNetwork when constructing the network for each sample.

plot_title

Passed to buildRepSeqNetwork when constructing the network for each sample.

color_nodes_by

Passed to buildRepSeqNetwork when constructing the network for each sample.

output_dir

The file path of the directory for saving the output. The directory will be created if it does not already exist.

output_type

A character string specifying the file format to use for saving the output. Valid options include "csv", "rds" and "rda".

output_dir_unfiltered

An optional directory for saving the unfiltered network data for each sample. By default, only the filtered results are saved.

output_type_unfiltered

A character string specifying the file format to use for saving the unfiltered network data for each sample. Only applicable if output_dir_unfiltered is non-null. Passed to buildRepSeqNetwork when constructing the network for each sample.

...

Other arguments to buildRepSeqNetwork when constructing the network for each sample, not including node_stats, stats_to_include or cluster_stats (see details).

Details

All node-level and cluster-level network properties are automatically computed for each sample's network. The node-level properties are renamed to reflect their correspondence to the sample-level network. Specifically, the properties are named ClusterIDInSample, SampleLevelNetworkDegree, SampleLevelTransitivity, SampleLevelCloseness, SampleLevelCentralityByCloseness, SampleLevelCentralityByEigen, SampleLevelEigenCentrality, SampleLevelBetweenness, SampleLevelCentralityByBetweenness, SampleLevelAuthorityScore, SampleLevelCoreness, SampleLevelPageRank. A variable SampleID is added to both the node-level and cluster-level meta data for each sample.

After the clusters in each sample are filtered, the node-level and cluster-level metadata are saved in the respective subdirectories node_meta_data and cluster_meta_data of the output directory specified by output_dir; each is saved using the sample ID (the corresponding element of sample_id_list) as a filename prefix, followed by the appropriate file extension based on output_type.

The unfiltered network results for each sample can also be saved by supplying a directory to output_dir_unfiltered, if these results are desired for downstream analysis.

The files containing the node-level metadata for the filtered clusters can be supplied to buildPublicClusterNetwork in order to construct a global network of public clusters. See the Searching for Public TCR/BCR Clusters package vignette for more details. The vignette can be accessed offline using vignette("public_clusters").

If the full global network is too large to practically construct, the files containing the cluster-level meta data for the filtered clusters can be supplied to buildPublicClusterNetworkByRepresentative to build a global network using only a single representative sequence from each cluster. This allows prominent public clusters to still be identified.

Value

A list of network objects as returned by buildRepSeqNetwork.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

Webpage for the NAIR package

Searching for Public TCR/BCR Clusters vignette

Author

Brian Neal (Brian.Neal@ucsf.edu)

See also