Find Public Clusters Among RepSeq Samples

Given multiple samples of RepSeq data, perform network analysis on each sample individually and filter clusters by node count and clone count.

Usage

findPublicClusters(

  ## Input ##
  file_list, input_type,
  data_symbols = NULL, header = TRUE, sep = "",
  sample_ids = 1:length(file_list),
  seq_col, count_col = NULL,
  min_seq_length = 3, drop_matches = "[*|_]",

  ## Network ##
  top_n_clusters = 20,
  min_node_count = 10,
  min_clone_count = 100,

  ## Visualization ##
  plots = FALSE, print_plots = FALSE,
  plot_title = "auto", color_nodes_by = "cluster_id",

  ## Output ##
  output_dir = file.path(getwd(), "public_clusters"),
  output_type = "rds",
  output_dir_unfiltered = NULL,
  output_type_unfiltered = "rds",
  ...

)

Arguments

file_list

Passed to loadDataFromFileList(one data frame per sample).

input_type

Passed to loadDataFromFileList.

data_symbols

Passed to loadDataFromFileList.

header

Passed to loadDataFromFileList.

sep

Passed to loadDataFromFileList.

sample_ids

A character or numeric vector of sample IDs, whose length matches that of file_list. The values should be valid for use in filenames.

seq_col

The column name or number of each sample's data frame that contains the receptor sequences to be used as the basis of similarity between rows during network analysis. This column must have the same name in each sample.

count_col

The column name or number of each sample's data frame that contains the measure of clonal abundance to be used when filtering the clusters. This column must have the same name in each sample.

min_seq_length

Passed to buildRepSeqNetwork when performing network analysis on each sample (prior to filtering the clusters).

drop_matches

Passed to buildRepSeqNetwork.

top_n_clusters

The top_n_clusters clusters with the highest node count will automatically be included among the public clusters.

min_node_count

Clusters with node count of at least this value will be included among the public clusters.

min_clone_count

Clusters with aggregate clone count of at least this value will be included among the public clusters.

plots

Passed to buildRepSeqNetwork when performing network analysis on each sample (prior to filtering the clusters).

print_plots

Passed to buildRepSeqNetwork.

plot_title

Passed to buildRepSeqNetwork.

color_nodes_by

Passed to buildRepSeqNetwork.

output_dir

The outpout directory to save the filtered network results for each sample.

output_type

A character string specifying the file format used to save the filtered results. Options include "rds", "csv" and "rda".

output_dir_unfiltered

The output directory to save the unfiltered network results for each sample, if desired. If NULL (the default), only the filtered results are saved.

output_type_unfiltered

A character string specifying the file format scheme used to save the unfiltered network results for each sample. Only applicable if output_dir_unfiltered is non-null. Passed to buildRepSeqNetwork.

...

Other arguments to buildRepSeqNetwork(), not including node_stats, stats_to_include or cluster_stats (see details).

Details

All node-level and cluster-level network properties are automatically computed for each sample. The node-level properties are renamed to reflect their association to the sample-level network. Specifically, the properties are named ClusterIDInSample, SampleLevelNetworkDegree, SampleLevelTransitivity, SampleLevelCloseness, SampleLevelCentralityByCloseness, SampleLevelCentralityByEigen, SampleLevelEigenCentrality, SampleLevelBetweenness, SampleLevelCentralityByBetweenness, SampleLevelAuthorityScore, SampleLevelCoreness, SampleLevelPageRank. A variable SampleID is added to both the node-level and cluster-level meta data for each sample.

After the clusters in each sample are filtered, the node-level and cluster-level meta data are saved in the subdirectories node_meta_data and cluster_meta_data, respectively, of output_dir; each is saved using the sample ID (the corresponding element of sample_id_list) as a filename prefix, followed by the appropriate file extension according to output_type.

The files containing the node-level meta data for the filtered clusters can be supplied to buildPublicClusterNetwork in order to combine the public clusters across samples and perform network analysis. The files containing the cluster-level meta data for the filtered clusters can be supplied to buildPublicClusterNetworkByRepresentative to build a network using only a single representative sequence from each cluster.

The unfiltered network results for each sample can also be saved by supplying a directory to output_dir_unfiltered, if these results are desired for downstream analysis.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

https://www.frontiersin.org/articles/10.3389/fimmu.2023.1181825

https://github.com/mlizhangx/Network-Analysis-for-Repertoire-Sequencing-

Author

Brian Neal (Brian.Neal@ucsf.edu) buildPublicClusterNetworkbuildPublicClusterNetworkByRepresentativePublic Clusters Vignette