Compute Cluster-Level Network Statistics

Given immune repertoire sequence data and the adjacency matrix for the corresponding repertoire network, perform clustering and compute cluster-level network statistics.

Usage

getClusterStats(data,
                adjacency_matrix,
                seq_col = NULL,
                count_col = NULL,
                cluster_id_col = NULL,
                degree_col = NULL,
                cluster_fun = cluster_fast_greedy)

Arguments

data

A data frame containing the immune repertoire sequence data, with clones/cells indexed by row.

adjacency_matrix

A square adjacency matrix corresponding to the repertoire network for the rows in data, with row dimension matching that of data. Sparse matrices created using the Matrix package are supported.

seq_col

Optional. The column name or number of data containing the receptor sequences (e.g., TCR CDR3 nucleotide sequence or amino acid sequence). If supplied, then related cluster-level properties will be computed.

count_col

Optional. The column name or number of data containing the counts (i.e., clone count or UMI count). If supplied, then related cluster-level properties will be computed.

cluster_id_col

The column name or number of data containing the cluster membership for each node. If NULL, clustering will be performed using igraph::cluster_fast_greedy() in order to determine cluster membership.

degree_col

The column name or number of data containing the network degree for each node. If NULL, the network degree will be computed using igraph::degree.

cluster_fun

Passed to addClusterMembership. Controls the algorithm used for cluster identification.

Details

The cluster-level statistics diameter_length, assortativity, transitivity, edge_density, degree_centrality_index, closeness_centrality_index, eigen_centrality_index, and eigen_centrality_eigenvalue are computed for each cluster by first generating the network graph for the cluster using the adjacency submatrix corresponding to that cluster. As a result, computing cluster statistics can potentially take some time for very large networks with many large clusters.

Value

A data frame containing one row for each cluster and the following columns:

cluster_id

The cluster ID number.

node_count

The number of nodes in the cluster.

mean_seq_length

The mean sequence length in the cluster.

mean_degree

The mean network degree in the cluster.

max_degree

The maximum network degree in the cluster.

seq_w_max_degree

The receptor sequence possessing the maximum degree within the cluster.

agg_count

The aggregate count among all nodes in the cluster (based on the counts in count_col).

max_count

The maximum count among all nodes in the cluster (based on the counts in count_col).

seq_w_max_count

The receptor sequence possessing the maximum count within the cluster.

diameter_length

The longest geodesic distance in the cluster, computed using length(igraph::get_diameter()).

assortativity

The assortativity coefficient of the cluster's graph, based on the degree (minus one) of each node in the cluster (with the degree computed based only upon the nodes within the cluster). Computed using igraph::assortativity_degree.

global_transitivity

The transitivity (i.e., clustering coefficient) for the cluster's graph, which estimates the probability that adjacent vertices are connected. Computed using igraph::transitivty() with type = "global".

edge_density

The number of edges in the cluster as a fraction of the maximum possible number of edges. Computed using igraph::edge_density().

degree_centrality_index

The cluster-level centrality index based on degree within the cluster graph. Computed using igraph::centr_degree()$centralization.

closeness_centrality_index

The cluster-level centrality index based on closeness, i.e., distance to other nodes in the cluster. Computed using igraph::centr_clo()$centralization.

eigen_centrality_index

The cluster-level centrality index based on the eigenvector centrality scores, i.e., values of the first eigenvector of the adjacency matrix for the cluster. Computed using igraph::centr_eigen()$centralization.

eigen_centrality_eigenvalue

The eigenvalue corresponding to the first eigenvector of the adjacency matrix for the cluster. Computed using igraph::eigen_centrality()$value.

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

if (FALSE) {
# Simulate some toy data for demonstration
toy_data <- simulateToyData()

# Generate network for data
net <- generateNetworkObjects(toy_data, "CloneSeq")

# Get cluster stats
cluster_info <-
  getClusterStats(net$node_data, net$adjacency_matrix,
                  seq_col = "CloneSeq", count_col = "CloneCount")
}