Compute Cluster-Level Network Statistics

Given the node-level metadata and adjacency matrix for an immune repertoire network, perform clustering and compute cluster-level network properties.

Usage

getClusterStats(
    data,
    adjacency_matrix,
    seq_col = NULL,
    count_col = NULL,
    cluster_id_col = NULL,
    degree_col = NULL,
    cluster_fun = cluster_fast_greedy
)

Arguments

data

A data frame containing the node-level metadata for the network, with each row corresponding to a network node.

adjacency_matrix

The adjacency matrix for the network.

seq_col

Specifies the column of data containing the receptor sequences. Accepts either the column name or column index. If provided, then related cluster-level properties will be computed.

count_col

Specifies the column of data containing a variable with counts (such as clone count or UMI count). Accepts either the column name or column index. If provided, then related cluster-level properties will be computed.

cluster_id_col

If cluster analysis has been previously performed to partition the graph into clusters, this argument specifies the column of data containing the ID of the cluster to which each node belongs. Accepts either the column name or column index. If NULL, then addClusterMembership will be used to partition the graph into clusters.

degree_col

Specifies the column of data containing the network degree of each node. Accepts either the column name or column index. If NULL, the network degree of each node will be computed.

cluster_fun

If cluster_id_col = NULL, this argument specifies the clustering algorithm used to partition the graph into clusters. Passed to addClusterMembership.

Details

The cluster-level properties diameter_length, assortativity, transitivity, edge_density, degree_centrality_index, closeness_centrality_index, eigen_centrality_index, and eigen_centrality_eigenvalue are computed for each cluster using the corresponding submatrix of the network adjacency matrix.

Value

A data frame containing one row for each cluster and the following variables:

cluster_id

The cluster ID number.

node_count

The number of nodes in the cluster.

mean_seq_length

The mean sequence length in the cluster.

mean_degree

The mean network degree in the cluster.

max_degree

The maximum network degree in the cluster.

seq_w_max_degree

The receptor sequence possessing the maximum degree within the cluster.

agg_count

The aggregate count among all nodes in the cluster (based on the counts in count_col).

max_count

The maximum count among all nodes in the cluster (based on the counts in count_col).

seq_w_max_count

The receptor sequence possessing the maximum count within the cluster.

diameter_length

The longest geodesic distance in the cluster, computed as the length of the vector returned by get_diameter.

assortativity

The assortativity coefficient of the cluster's graph, based on the degree (minus one) of each node in the cluster (with the degree computed based only upon the nodes within the cluster). Computed using assortativity_degree.

global_transitivity

The transitivity (i.e., clustering coefficient) for the cluster's graph, which estimates the probability that adjacent vertices are connected. Computed using transitivity with type = "global".

edge_density

The number of edges in the cluster as a fraction of the maximum possible number of edges. Computed using edge_density.

degree_centrality_index

The cluster-level centrality index based on degree within the cluster graph. Computed as the centralization element of the output from centr_degree.

closeness_centrality_index

The cluster-level centrality index based on closeness, i.e., distance to other nodes in the cluster. Computed using centralization.

eigen_centrality_index

The cluster-level centrality index based on the eigenvector centrality scores, i.e., values of the first eigenvector of the adjacency matrix for the cluster. Computed as the centralization element of the output from centr_eigen.

eigen_centrality_eigenvalue

The eigenvalue corresponding to the first eigenvector of the adjacency matrix for the cluster. Computed as the value element of the output from eigen_centrality.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

Webpage for the NAIR package

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

if (FALSE) {
# Simulate some toy data for demonstration
toy_data <- simulateToyData()

# Generate network for data
net <- generateNetworkObjects(toy_data, "CloneSeq")

# Get cluster stats
cluster_info <-
  getClusterStats(net$node_data, net$adjacency_matrix,
                  seq_col = "CloneSeq", count_col = "CloneCount")
}