Compute the Neighborhood Around a Cell/Clone

Given immune repertoire sequence data and a target receptor sequence, return the subset of the data consiting of those cells/clones whose receptor sequences are within a specified distance of the target.

Usage

getNeighborhood(data, seq_col, target_seq,
                dist_type = "hamming", max_dist = 1)

Arguments

data

Data frame containing the RepSeq data.

seq_col

The column name or number of data containing the receptor sequences of the same type as target_seq.

target_seq

A character string containing the receptor sequence of the cell/clone around which to form a neighborhood.

dist_type

A character string specifying the distance type on which the neighborhood is based; valid options are "hamming" and "levenshtein".

max_dist

A numeric scalar specifying the neighborhood radius. The neighborhood consists of all clones whose receptor sequences are at most max_dist from the target sequence.

Details

The neighborhood will only be computed if target_seq is present in the data, i.e., in the column of data referenced by seq_col.

Value

If target_seq is present in the data, a data frame containing the subset corresponding to the neighborhood; otherwise, the function returns NULL.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. https://doi.org/10.3389/fimmu.2023.1181825

https://www.frontiersin.org/articles/10.3389/fimmu.2023.1181825

https://github.com/mlizhangx/Network-Analysis-for-Repertoire-Sequencing-

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

# Simulate some toy data for demonstration
toy_data <- simulateToyData(sample_size = 500)

# Get neighborhood around first clone sequence
nbd <- getNeighborhood(toy_data, seq_col = "CloneSeq",
                       target_seq = toy_data$CloneSeq[[1]])