Save List of Network Objects

Given a list of network objects such as that returned by buildRepSeqNetwork or generateNetworkObjects, saves its contents according to the specified file format scheme.

Usage

saveNetwork(
  net,
  output_dir = getwd(),
  output_type = "individual",
  output_filename = "MyRepSeqNetwork",
  pdf_width = 12,
  pdf_height = 10
)

Arguments

net

A list conforming to the format returned by buildRepSeqNetwork or generateNetworkObjects. See details.

output_dir

A file path specifying the directory in which to write the file(s).

output_type

A character string specifying the file format scheme to use when writing output to file. Valid options are "individual", "rds" and "rda". See detials.

output_filename

A character string. All files saved will have file names beginning with this value.

pdf_width

If the list contains plots, this controls the width of each plot when writing to pdf. Passed to the width argument of the pdf function.

pdf_height

If the list contains plots, this controls the height of each plot when writing to pdf. Passed to the height argument of the pdf function.

Details

The list provided to the net argument must contain at least the elements igraph (of class igraph), adjacency_matrix (of class matrix or dgCMatrix), and node_data (a data frame). The components cluster_data (a data frame) and plots (a list of ggraph objects) are optional, and will also be saved if present.

By default, each list element is saved as a separate, uncompressed file: data frames, including node_data and cluster_data, are saved as .csv files; igraph is saved as a .txt (using write_graph with format = "edgelist"); adjacency_matrix is saved as a .mtx file (for the dgCMatrix class) or .csv file (for the Matrix class); the elements of plots are printed to a .pdf file using pdf, with the pdf file containing one plot per page.

Setting output_type to "rda" or "rds" writes the entire list net to a single compressed file (a .RData file or a .RDS file, respectively), while also printing the plots to a .pdf file.

References

Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. Frontiers in Immunology, vol. 14. doi: 10.3389/fimmu.2023.1181825

Webpage for the NAIR package

Author

Brian Neal (Brian.Neal@ucsf.edu)

Examples

toy_data <- simulateToyData()

net <-
  generateNetworkObjects(
    toy_data,
    "CloneSeq"
  )
#> Computing network edges based on a max hamming distance of 1... Done.
#> Network contains 122 nodes (after removing isolated nodes).

saveNetwork(
  net,
  output_dir = tempdir()
)
#> Node-level meta-data saved to file:
#>   /tmp/RtmpaCqR7q/MyRepSeqNetwork_NodeMetadata.csv
#> Network igraph saved in edgelist format to file:
#>   /tmp/RtmpaCqR7q/MyRepSeqNetwork_EdgeList.txt
#> Adjacency matrix saved to file:
#>   /tmp/RtmpaCqR7q/MyRepSeqNetwork_AdjacencyMatrix.mtx

# clean up temp files
file.remove(
  file.path(
    tempdir(),
    c("MyRepSeqNetwork_NodeMetadata.csv",
      "MyRepSeqNetwork_EdgeList.txt",
      "MyRepSeqNetwork_AdjacencyMatrix.mtx"
    )
  )
)
#> [1] TRUE TRUE TRUE